Fixed support to remove the enclosing method of a type while obfuscating.
Previous support was not complete since it was not working when shrobbing and
dex generation were not done in the same plan.
This support removes the information from the IR, so after the obfuscation
phase, if the flags do not explicitely keep it, the enclosing method of a
typecannot be retrieved.

Bug: 11803482

Change-Id: I36d6b02a54197c8dd3dc93ca81b607a66718c61a