Fixed support to remove the enclosing type of a type while obfuscating.
Previous support was not complete since it was not working when shrobbing and
dex generation were not done in the same plan.
This support removes the information from the IR, so after the obfuscation
phase, if the flags do not explicitely keep it, the enclosing type of a
type cannot be retrieved.

Bug: 11803482

Change-Id: I4ff69d47c60f2406cf18d8e740410b7596e2474e