Fixed support to remove defaut value in annotations while obfuscating.
Previous support was not complete since it was not working when shrobbing and
dex generation were not done in the same plan.
This support removes the information from the IR, so after the obfuscation
phase, if the flags do not explicitely keep it, the default value of an
annotation cannot be retrieved.

Bug: 11803482

Change-Id: If9b3966f8bc23093b09714e427ec26c4a130a99d