Fixed support to remove the thrown exceptions of a method while obfuscating.
Previous support was not complete since it was not working when shrobbing and
dex generation were not done in the same plan.
This support removes the information from the IR, so after the obfuscation
phase, if the flags do not explicitely keep it, the thrown exceptions of a
method cannot be retrieved.

Bug: 11803482
Change-Id: I78efd58fd1b1b0f96ff1df5a8b046c23f1477c93