Removed wrong assertion in ReflectAnnotationsAdder
There are two different ways of ensuring generic signatures for JMethod and
JField, one is an assertion the other is an 'if'.
A previous patch (I174dc97615ee61817e12e6a6b6351ad78838fbbe) wrongly changed that.
Note that this will be obsolete when the marker OriginalTypeInfo will be
splitted in two.

Change-Id: Ia80cd128e5dd6bcb79b4bb96add4a8251e8b7990