Fixed ClassCastException when the classpath is not complete.
Bug: 12696838
When using the shrobbing feature, we used to assume that the entire hierarchy
of the types to emit was defined, so we used to cast directly types to
the defined types which resulted in ClassCastexception when it wasn't the case.
This patch modifies the member finder used by the literal refiners so that a
phantom class in hierarchy will not cause an exception.

Change-Id: I6737e4b70121833659e0cb60ae0d84b5da43ea17