Fix redirection in DeviceRunner
Redirection requested by user is only active for the test
which is executed, and not all the device setup commands.
Thus tests that rely on output comparison are more likely
to work since not polluted by extra output.

I also fixed a path issue due the change done in test
properties for jack.home.

Change-Id: I2c7d758946d03ac3efa7db08566bd37ae98b373e