Stop creating message in NonSuchFileException unnecessarily
NoSuchFileException is thrown 29686 times during compilation of Music app
but its message is not obtained even once. location.getDescription is not
a cheap operation and message creation uses string concatenation as well.
Instead of creating a message eagerily this change stores a reference to
a location in the NonSuchFileExpection object and creates the message
only when it's needed. This speeds up the creation of NoSuchFileException
almost 40 times which makes a difference if we multiply it by almost 30k times
it's created. We will also save on GC as it saves at least 3 object creations
per NoSuchFileException which makes over 90k object creations saved.

Fix also other exceptions in com.android.sched.util.file.

Change-Id: Ibc2a413891a2e24be786f440196e9a4276215cc0