Optimize VPathFragment creation.
When compiling Music app we create 207k instances of VPathFragment.
83k of them are created in VPath.appendPath and VPath.prependPath.
Those instances are the same so this change introduces a new static
field to VPath which will keep a singleton for all those invocations.
The remaining 124k instances are created in VPath constructor and have
the same separator (INTERNAL_SEPARATOR). This change introduces a new
subclass of VPathFragment to handle separators different than
INTERNAL_SEPARATOR. VPathFragment itself assumes INTERNAL_SEPARATOR is
used. This way we make those 124k objects smaller and save on GC.

Change-Id: Ic69c6c2bcf72473ea71a2d7d0adc1bfb5dc5bbd7