Reduce the number of JayceReader constructor created.
In JayceReaderFactory.get we use a Constructor<?> of a JayceReader.
This reader is always the same for a library but currently we obtain
the Constructor<?> object for each type in a library. This change
keeps the Constructor<?> object for each JackInputLibrary in the
library object. This way it can be created only once per library and
reused for reading all the types in the library.

Change-Id: I2786624b8b51339645d16bfc71f5773d2ec3da37