Read some sections of dex in DexBuffer constructor
I profiled the incremental compilation of Music app and it showed
that we are reading and creating 1,330,524 strings from dex files
in DexBuffers. I checked and the total of strings in all dex files
represented by DexBuffers is 284979. That's around 20% of what we
were reading. The conclusion was that we read the same strings
over and over again. Then I looked at ConstantManager.addDexFile
and it striked me that we need all strings, typeIds, typeNames,
fieldIds and methodIds anyway because we iterate over them in
addDexFile. This means that we can as well read all of them
when DexBuffer is created. This way we save a lot of cpu and
memory allocations. Strings alone were taking 70MB of heap.
Now we're allocating only 20% of it.

This change speeds up incremental compilation by over 12.5%.

Change-Id: Iee18e9d17419886e552e3fd0b1143d2bb7220c9c