Avoid comparing full signatures of methods.
TypePackageAndMethodFormatter.getName(JMethod) is not cheap both
cpu and memory wise. In ReferenceMapper.findMethod we don't need
to compare MethodBindings with signature of every method. Instead
we can first compare the name of the method and the number of
parameters and only when those are equal check the full signature.

This saves a lot of memory allocations. Before findMethod was making
1,215,894 allocations for over 77MB of memory. After this change it
makes only 84,391 allocations for 5MB of memory. This reduces the total
amount of memory used during compilation
(org.eclipse.jdt.internal.compiler.ProcessTaskManager.run) from 951MB
to 879MB which is over 7.5%.

Change-Id: I01362d881e9d64ee7a0b6d06e83558dae17c1b87