Avoid StringBuffer resizing in ReferenceMapper.signature
While profiling I found out that ReferenceMapper.signature for both
methods and fields makes 1,214,784 allocations for over 143MB of memory.
Further investigation showed that big part of them was because of
StringBuffer having to extend its capacity. We can prevent it by
calculating the total number of characters in a resulting string and
creating a StringBuffer with exactly that size. We can do it cheaply
because '.signature()' methods in *Bindings are cached and each invocation
after the first one is just returns already calculated char[].

This saves a lot of memory allocations. Before signature methods were making
1,214,784 allocations for over 143MB of memory. After this change they
make only 784,567 allocations for 86MB of memory. This reduces the total
amount of memory used during compilation
(org.eclipse.jdt.internal.compiler.ProcessTaskManager.run) from 951MB
to 893MB which is over 6%.

Change-Id: I818dc62877369505a8698912aef49585df2b7779