Reduce the memory usage in ReferenceMapper
ReferenceMapper.get(FieldBinding) and ReferenceMapper.get(MethodBinding)
create a signature of a binding to use it as a key in |methods| and
|fields| maps. This requires creation of a string which is a copy of
data from binding. We can avoid this and reduce the ammount of memory
used by using a special SignatureKey instead of string signature.
The new key shares data with binding so we don't need to copy it.

Before this change ReferenceMapper.get(FieldBinding) was allocating
65MB of memory with 957,507 allocations (1% of all allocated memory) and
ReferenceMapper.get(MethodBinding) was allocating 71MB of memory with
1,138,383 allocations (1,1% of all allocated memory) during Music app
compilation.

After this change ReferenceMapper.get(FieldBinding) allocates 1MB of
memory with 32,666 allocations (0.0% of all allocated memory) and
ReferenceMapper.get(MethodBinding) allocates 10MB of memory with
231,556 allocations (0.2% of all allocated memory). This means we
saved 125MB of memory and 1,831,668 allocations (1.9% of all allocated
memory).

Change-Id: I91f08a7ac13271d67803028955156f4afdc5d54f