Reduce the number of iterators created.
Profiler showed that Collection.iterator() allocates 6% of
all allocated memory. This is 406MB for Music app. It turns
out that big part of this is called in JVisitor.accept(Collection<T>).
At least half of the time we know that the collection is an ArrayList
and we can iterate through it without an iterator by just calling
ArrayList.get(index).

This change introduces new accept(ArrayList<T>) method to JVisitor and
JVisitorWithConcurrentModification. Those methods iterate over the
collection without iterator. Also in JVisitorWithConcurrentModification
when we know that copy is an ArrayList we don't need to use iterator.

This change reduces the ammout of memory allocated by
Collection.iterator() to 3% from 6%. For Music app this is 220MB instead
of 406MB. It also means over 5,000,000 allocations less.
Impact on performance is also positive as iterator makes more operations
than get() method.

Change-Id: I7e55a32b5ff01eb9b071908f1de5e21a1fb70f00