Make ControlFlowGraph immutable.
Decouple BasicBlock from ControlFlowGraph. It's a cyclic dependency
which is never good. Collect all blocks in CfgBuilder.BuilderVisitor
and use them to build immutable ControlFlowGraph at the end.

Graph.getNodes used to clone nodes. This was allocating 1.6% of all
allocated memory. For Music app that was 89M and 3,714,520 allocations.
It was also a CPU intensive and was taking 1.2% total CPU.

This change reduces the total ammount of
memory allocated by 2.0% which for a Music app is 106MB.

Change-Id: I835a297d8f812d130c3bded3cd6f17f35072be62