Allow to use two enums with the same simple name into a class
- To support partial recompilation with enum used into a switch statement,
Jack generates a field and a method allowing to get a mapping between
an enum field and an integer that will be used into a switch statement.
This generated code must not take only the simple name of enumeration as
discriminant, since several enum can have the same name. Thus, it is
better to use his fully qualified name.

Bug: 19639131
Change-Id: I2dbeef61e24e0e7b7250d3f0cfcbb619bd6e7829