Fix bugs when using incremental with --output-jack
- When --output-jack is used, incremental folder contains all dex files
and jayces files of imported libraries, thus Jack must remove imported
libraries and import only incremental folder when --output-jack is used
to avoid type conflict. It seems more performant than using collision policy.
- Since ReadWriteZipFS and VFStoVSTWrapper is used incremental support with
--output-jack was broken because input jack library has the location of
incremental folder and output jack library has the location of the target
jack file. Thus comparing location to avoid to copy dex file from a library
to the same library does not work. I clean library usages to always used
ReadWriteZipFS when it is required and to be able to properly skip dex copy
when source and target library are the same library.

Bug: 21600202

Change-Id: I0c5e899e12033f608d43cae9baecc1227ea158c9