Add optimized enum switch support.
1. Develop the optimized switch enum support. It could
be used to improve the current switch enum support.

2. Add test cases for optimized enum switch support.
Right now the test cases cover the cases where enum is
defined and used in application, enum is declared in
framework and used in application. Tests also cover the
situations in which enum is declared as: public, private,
protected and package visible.

3. Add strategies code to determine whether it is worth
effort to optimize the code.

Change-Id: I70ec82d05ef14971e31cb497e9308eafbef086d6