Remove cleanup at onSystemStart
Rely on File.deleteOnExit for deleting files even when server exists
before GC has collected their corresponding ClassLoader.
Add a new command to jack-admin: server-cleanup. This command is to
run while server if off and allows to clean file that could be
forgotten on disk in case of VM crash after an update.

Change-Id: Ia8552a0098a02951d14c2190946ddf494c9efc4c