Fix runtime illegal access error
- Add inner test that trigger an illegal access error at runtime
when calling the constructor of B if markers of method call are
not cloned.
- Update CloneExpressionVisitor to clone markers of expressions as
it was done for statements. If markers of JMethodCall are not cloned,
the InnerAccessorGenerator does not work correctly and forgot to
generate bridge which are required to avoid illegal access error.

Change-Id: I2aa1ee3bc8923d4a7bff102ba37f2a232833b10b

(cherry picked from commit 90520ea1a2bffccbe5380386d9292c715631502f)