Update InnerAccessorGenerator with default method
- InnerAccessorGenerator must take into account super call to
default method to generate accessor in the right place. In case
of non default method, accessor is generated into the first enclosing
class that extends the type containing the called method. In case
of a default method, accessor is generated into the first enclosing
class that implements the type containing the called method.

Change-Id: Idd5b3c5e312d3d44848258b62165cc11e6bcbfde