Modified MappingApplier to avoid name collision.
The MappingApplier is now able to search a class
hierarchy to avoid renaming two member with the
same name.
+ Added shrob test which create a name collision.
The test contains a class and its subclasses containing
both a method with the same signature and a mapping
forcing them to the same name.
It doesn't avoid collision on types yet.

Bug: 25001317
Change-Id: Ib1f52d1db5d35c751f39af8a1b54d0e9b9dc9bbb