Check for triggers of a Dalvik bug with Arrays
The bug is triggered by
- check-cast
- instance-of
- *new-array
- const-class
when they target an array of a protected inner class from a class that
should have access but is not in the same package. The bug is present in
Dalvik K, not in Art.

Run the checker when running with min target API is 19 (KitKat) or less.
When triggering code is found, warn about about the bug.

Bug: https://code.google.com/p/android/issues/detail?id=178175

Change-Id: If61c52ff4178e8a1becf50ff3e0d565727cf5f13