Code coverage report tool
Adds jack-jacoco-report-tool project. It is a code coverage
reporting tool based on Jacoco libraries and dedicated to Jack.

While Jacoco relies on original class files to analyze coverage
execution file, this tool works on a metadata JSON file produced by
Jack. The tool is a kind of extension and acts as a frontend to the
exisiting Jacoco report library.

Bug: 20115492
Change-Id: I7e35fe6463f3e984d3b2f7120210133c6a4b97b5