Fix JarjarTests.jarjar004() by specifying not to zip dex
This broke regression server.

This is due to a bug in api based toolchains which checked
whether the output File was a directory or not, and set
Jack option accordingly whereas it shouldn't, as
it is done in cli based toolchain.

This issue is addressed in another CL still in review
at the moment.
(https://android-review.googlesource.com/#/c/181957/)

Bug: 26486911

Change-Id: Id94308c87e91d32ae4e4d6673e9820d5507cd684