Code coverage: support class filtering
Adds properties jack.coverage.jacoco.include and
jack.coverage.jacoco.exclude that respectively define which classes
must be included or excluded from code coverage instrumentation.

They are implemented with the CoverageFilterSet class which is parsed
from the properties using the CoverageFilterSetCodec. Both include
and exclude CoverageFilterSet objects are managed in a CoverageFilter
singleton object which indicates whether the class must be
instrumented for code coverage. By default, classes from org.jacoco.*
(Jacoco) and com.vladium.emma.* (Emma) are excluded so we do not
instrument the classes of the imported instrumentation library.

Bug: 26864099
Change-Id: I6540f310199c35d2abe8543e9fe12504f3d34afd