Code coverage: support imported libraries
Visits all classes that will be emitted in the final DEX file.

Adds the coverage enabling property to the DumpInLibrary
category so pre-dex is automatically disabled if there is a mismatch.

Adds @Protect constaint on JDefinedClassOrInterface to ensure the
CodeCoverageSelector do not miss classes that may be generated during
the compilation (like synthetic classes).

Bug: 20115492
Bug: 26864099
Change-Id: I5045cc309cad263e999b9dbf6287c5fc934cc115