Cleanup code coverage
Analyzes probes of a method in one pass (instead of three) by
visiting blocks in depth-first order. The blocks that terminate a
probe will be marked with a ProbeMarker. This marker is used by the
instrumentation to know which blocks must be instrumented

Also adds a new coverage test with exception flow and fixes a couple
issues discovered by FindBugs.

Change-Id: Ifada12e5e50631dc9116c7f7eb68800da1ad31c5