Coverage: find runtime package if it is not provided
We used to expect that the internal Jacoco package name is provided
through property 'jack.coverage.jacoco.package' since that package
changes for each Jacoco release. Since it is not suitable for SDK
integration, this CL provides a fallback mode where we look for the
org.jacoco.agent.rt.internal* package (where '*' can be any
suffix) if the 'jack.coverage.jacoco.package' property is not
provided.

This CL also caches references to the Jacoco class and method used
for instrumentation to avoid looking it up for each instrumented
class.

Bug: 27237961
Change-Id: I31ae876ae6d3b8d2fe15077355a3bfa9d7b94650