Replace <symlink> of version.properties by <copy>
<symlink> is not portable to Windows.

This CL forces a copy of the version.properties file to Jack and Jill
resources to be able to run tests in Eclipse.

One asset of using a symlink was the coherency of content for all those
files.

In an attempt to keep that, I added commands to prevent modification of
content on these copies by making them read only with <chmod> for unix-styleOS
and <attrib> for Windows. These task have simply no effect
if not executed on the os family that supports them. This way the code remains
quite simple, it does not introduce Ant conditional tests about the OS family.

Bug: 27606308

Change-Id: I74e9d2ee04004157eda43d5322b18550dbfee6f4