Modify Reaching definition marker to support IR modification
- Add a test that trigger the problem when using jill toolchain
- DefUsesChainsSimplifier commits IR transformations on the fly, but these
IR transformations invalidate reaching definitions during the optimization
which can lead to apply optimization when it is not allowed.
- Rather than computing the variable that reach a block by reading the
expression that defines it, keep the variable directly into the marker such as
even if the expression is previously optimized the information is not lost.
- Remove ReachingDefinitionsMarker directly from DefUsesChainsSimplifier since
the information is no longer valid after this optimization.

Bug: 27620766

Change-Id: I6b090436083dad9450a8abc4576bc4f3a2c9c9ed