Ignore deprecated references in Javadoc
Note that we do not have a project configuration for Javadoc warnings,
and I added this one rule by editing the file directly.
Adding just this one rule could not have been done by using the UI, so
maybe doing this is a bit hackish but I didn't want to make the
decision for all rules.

Change-Id: I41c26460666119a850d7faef697bc774b4b24a3d