Avoid closing the same VFS twice
This is to avoid closing twice the same VFS that is shared between an
input library and an output library in incremental mode.
This is pretty hackish but I think it is safe until we rework how we
close VFSes and libraries.
This will allow to put back the assertions on the "close" state in VFS.

Bug: 27717048
Change-Id: I63227cbb0f03726bb3e0bd214b1fb612337b2418