Add filter annotation to schedulables
Adds the @Filter annotation to schedulables which need to run in
a specific set of components.

This CL introduces two filters to replace previous "filtering"
adapters:
* SourceTypeFilter replaces ExcludeTypeFromLibAdapter
* TypeWithoutPrebuiltFilter replaces
  ExcludeTypeFromLibWithBinaryAdapter

Schedulables running on all classes are not annotated with the
@Filter so they run with the default adapter.

Note: this CL does not update the structure of the plan yet. This
will be done in a separate CL.

Change-Id: Ib979dd656e4facc5101f211def63971559f6127d