Code coverage: update class id computation
Computing the class ID from the class name only leads to conflict
when code coverage is applied to the platform. Also it does not take
into account the code modifications applied onto a class (like
shrinking).

This CL updates class ID computation to reduce chances of conflict:
- includes field name and type
- includes method name and signature
and moves it after the probe analysis so we are sure that we compute
the class ID *after* modifications happened.

Bug: 28034736
Change-Id: Ib35ff7cdbec90ba4d6374e01574c8bb11147d003