Fix shift const operand overflow issue
From the Javaspec:
https://docs.oracle.com/javase/specs/jls/se7/html/jls-15.html#jls-15.19
-----
If the promoted type of the left-hand operand is int, only the five
lowest-order bits of the right-hand operand are used as the shift
distance. It is as if the right-hand operand were subjected to a
bitwise logical AND operator & (§15.22.1) with the mask value 0x1f
(0b11111). The shift distance actually used is therefore always in the
range 0 to 31, inclusive.

If the promoted type of the left-hand operand is long, then only the
six lowest-order bits of the right-hand operand are used as the shift
distance. It is as if the right-hand operand were subjected to a
bitwise logical AND operator & (§15.22.1) with the mask value 0x3f
(0b111111). The shift distance actually used is therefore always in
the range 0 to 63, inclusive.
-----

Since Dalvik has no shift instructions that takes more than a 8 bit
distance, we are going to mask the distance according the spec when
generating ROP.

BUG: 28244190

Change-Id: Ifb3d95f0eddfd312f4b8920d6e9422073d861444