Orphan dex file must not always be merged
- With incremental support, orphan dex file can belong to the list of
type to emit, but also belong to the orphan dex file of an imported
library (it is the case for the incremental library) since output
and input are the same library. Nevertheless the dex file must be
merged only one time. We use a set to deduplicate the InputVFile
representing dex files that must be merged.

Bug: 28156189
Change-Id: Ifc6c0f3fa0815e1f9cd767f8b9543b54f447d17c