Orphan dex file must not always be merged
- With incremental support, orphan dex file can belong to the list of
type to emit, but also belong to the orphan dex file of an imported
library (it is the case for the incremental library) since output
and input are the same library. Nevertheless the dex file must be
merged only one time. We use a set to deduplicate the InputVFile
representing dex files that must be merged.

Bug: 28156189

(cherry picked from commit 2fe2d442f09c432d94659c278fbd31de853c1576)

Change-Id: I70fcf99c6c9442a1c2f8d7969d56970c7a053e8d