Let OutputZipFile handle streams instead of WriteZipFS
WriteZipFS used to open the stream in the constructor, which would clear
the remover and result in empty lib files when it is never written
because of an error.
It's better to let OutputZipFile handle the stream, so that is is opened
lazily.

Bug: 28329790

(cherry picked from commit a98a5b82224f9d7a30705f0daec321305503d905)

Change-Id: Idb5e1d0aa16522040fad1e60540608a0d77b42e9