Find bridges to add into lambda when coming from Jill
- When coming from Jill, it can happens that bridges needed by the lambda
are generated as default bridge into interfaces. Later into Jack, if classpath
is known, we check if there is default bridges in order to add it into the inner
classe representing the lambda. If classpath is unknow, Jack emits an error
because generated code can failed at runtime due to missing bridges on
older Android releases.

Bug: 28333827
Bug: 26765160
Change-Id: I8513d48be605cd27e9f70d2a458cd8133c69b9de