Remove redundant try/finally in BoostLockedRegion
Instead of conservatively insert try/finally around synchronized regions,
we can look at the monitor enter / exit points later in the code.

We know for sure that exit points must be part of a region that has a
catch-all handler so we can just insert the priority reset calls right
after it.

This speeds up AmSlam by another 10% bringing it to a total of 20%
improvement over no having BoostLockedRegionPriority.

Change-Id: I285fa6ac9b7c558e6d36208e9de2e696f2aafdbe