TypeLegalizer needs useless switches to be removed
- Java 7 allows to write switch on String, but it is
not supported by the type legalizer that assumes that switch
expression will be a primitive type that may need to be unbox.
Jack transforms switch on String into 'if' statement before
the type legalizer, but switches with only a default case are
not transformed because they are removed by the schedulable
'UselessSwitchesRemover'. Thus, this schedulable must be run
before the TypeLegalizer.

Bug: 28964547

Change-Id: If096df1669e44aa3eba16369d7dcc3f80ebcbf5c