Add an annotation for tests that need runtime prebuilts
The annotation is @Runtime.
This way the dumped test lists are able to specify
whether the test needs a runtime prebuilt, thus easing the job
of the regression server.
A new attribute has been created in the generated .js file listing
tests: runtimePostM the sole presence of which indicates a runtime test,
and the value of which determine wether this test is to be run
on runtimes starting from N.

Change-Id: I171dc7b0dc2f39b0fd9dfe3e408ce1f38fbe5d25