Add an annotation for tests that need runtime prebuilts
The annotation is @Runtime.
This way the dumped test lists are able to specify
whether the test needs a runtime prebuilt, thus easing the job
of the regression server.
A new attribute has been created in the generated .js file listing
tests: runtimePostM the sole presence of which indicates a runtime test,
and the value of which determine wether this test is to be run
on runtimes starting from N.

(cherry picked from commit bc1479b7e3f921ed6a94ea7abe4a1edbe49d392a)

Change-Id: Icd1ecc73728f4284ee198a9381403306db293c05