Revert "Revert "Add an annotation for tests that need runtime prebuilts""
This reverts commit 8e68cd039c53fe9218402a15ea637bbff8b3f6bf.

The annotation is @Runtime.
This way the dumped test lists are able to specify
whether the test needs a runtime prebuilt, thus easing the job
of the regression server.
A new attribute has been created in the generated .js file listing
tests: runtimePostM the sole presence of which indicates a runtime test,
and the value of which determine wether this test is to be run
on runtimes starting from N.

Change-Id: If15b81cdbe9c3bbbe4ab170cb8ecca95be0abf9a