Reduce thread contention on SourceInfoFactory
When multiple threads create source info elements from the factory,
they must synchronize the access to the different HashMaps. Each
access to the map is allowed to only one thread at a time. This is
causing thread contention.

This CL updates the SourceInfoFactory to use ConcurrentHashMap and
remove the need to synchronize. It greatly improves the read when
there is no need to update the map (because the requested element
already exists). Therefore it reduces thread contention.

Bug: 30216492
Change-Id: I05667b40d20e2ed27d5dbde9f6dc331cf2d74d33