Ensure that ConditionalBasicBlock have two targets
- Backend requires that ConditionalBasicBlock have two targets, one
for the then branch and one for the else branch.
- Ensure that after branch resolution, ConditionalBasicBlock have
two targets, before it can happens that the else target is null.
It happens only into dead code that is removed from the CFG, thus
backend is happy, nevertheless is prevent Jack to optimize
getSuccessors of ConditionalBasicBlock to avoid to filter null value.

Change-Id: I06a0940ad41d4ad2290d3bf0bb1f2434a0506495