Cfg-IR implementation: basic structure
This CL represent an initial design for cfg-IR basic
structures. It includes details of cfg-IR representation
as well as:
  (a) transformation building cfg-IR method body based on
      CFG marker, and
  (b) rewritten CodeItemBuilder so it builds CodeItem based
      on the information from cfg-IR, not on CFG marker.

Note that the representation is not a full implementation
of j-nodes, it only includes functionality needed to represent
cfg-IR and use it for building CodeItem.

All presubmit tests and test-long are passing.

Change-Id: I071790d21e0e6317c1349009eab5cacd5408efdf