Adding runtime test support to DexOutputBasedTest.
With this change DexOutputBasedTest will run runtime tests
on each call to compileAndValidate(...), potentially several
times per test to validate different compilation configurations.

Now DexOutputBasedTest runs the same tests each time, we might
consider to extend it in future to run different tests for
different compilation settings.

Now it is required to have at least a trivial runtime test,
this is in accordance with our previous discussions on that
matter.

Directory structure 'dx'/'jack' is used to match that of our
runtime infrastructure. The test(s) class must be located in
a directory 'dx' and have name Tests.

Change-Id: Ia96e77861e34f891b6e3a3779b7c6ef08fc72172