Update compatibility checking for API level
The AndroidCompatibilityChecker used to process the same methods that
are processed by the BridgeInInterfaceRemover. But after the change
561c9c08236c677472730127d302743b5b3ba779, this is no longer the case
because the AndroidCompatibilityChecker checks methods that are no
longer processed by the BridgeInInterfaceRemover. This change is
causing test failures.

Indeed the AndroidCompatibilityChecker used to be executed when
importing a library (--import) to a DEX file (--output-dex) and
check that the library is compatible with the sources being
compiled regarding a minimum Android API level. If both the library
and the DEX file use the same min API level (in both compilations),
it is not need to do the check twice if the library comes with DEX
prebuilt (which should be valid already).

This CL changes how we check the compatibility of Jack libraries
regarding a minimal Android API level:
* When a library is generated without DEX prebuilt, it is considered
agnostic of the min API level. Thus the compatibility check must
be done only when importing the library (to produce a DEX file).
* When a library is generated with DEX prebuilt, it is targeted for
a specific API level. Therefore we must do the compatibility
checking when generating the library.
* When importing a library, we can still enable/disable this
checking. However, the check is only necessary if the library has
no DEX prebuilt.

Bug: 30751382
Change-Id: I05a12cede297def027d29d5086d780044b470fbb