Reduce creation of dead basic block
- Do not longer create basic block a the end of JBlock. All blocks needed
by a JIfStatement are managed during the visit of JIfStatement rather than
in two place. No need to create new basic block at each end of JBlock
that have a catch list. End of virtual try block already had goto statement
to skip JCatchBlock. Block into a JLabeledStatement are managed during
the visit of JLabeledStatement.
- Create basic block of JCatchBlock in one place rather in than in two places,
it allows a better understanding.
- Checks that current statement is empty at the end of the cfg built.
- At the end of the cfg built all pending virtual statements are associated
to the exit block, to guarantee that after branch resolution all blocks
have targets
- Force all NormalBasicBlock to have a target even if there is no next
statement, in this case we target the exit block.
- This CL allows to reduce basic block created by about 10% when building
jack-wip-dev and it avoids to create about 95% of dead basic block.

Bug: 30821334

Change-Id: Ib724fd818f9d64c7755dbb1cd1b069ea880350f6