Remove entry block from nodes of CFG
- Entry and exit block of the CFG are blocks that are specifics.
Today, the nodes returned by the method getNodes of the CFG contains
the entry block but not the exit block, it is not homogeneous.
Now, nodes of the CFG does not longer contains the entry block,
nevertheless it can be retrieve by the method getEntryNode.
The nodes returned by the method getNodes return
only nodes that contains statement coming from the Jack IR.

Change-Id: I90b95d306d12e6faeec611e1a19be1f0a8d12c0a