Optimize branches in a reverse post DFS order.
Currently, if A -> {B, B} and B -> {C, C}, if the B branch is optimized
first, A branch can also be optimized but not the other way around.

With this change, we ensure a reverse post DFS Order which guarantees
B is optimized first so A would get a chance of being optimized.

We will no longer depend on the SsaMethod.getBlocks() ordering
which can be meaningless and slightly different if Jill is used.

BUG: 31026788

Change-Id: I6c5f4fd7194a4f8661b9186fa6e958c531278a3b