Make SsaBasicBlock's insns a List.
By making insns more data structure implementation independent,
we can build SsaBasicBlock using other type of list such as LinkedLists.

This is useful if we are trying to build SsaBasicBlocks directly from
Jack IR which involve inserting instructions in the beginning
of the list. Using ArrayList in that scenario is not efficient.

Change-Id: I148b14275ab9b5820863d1c61c434836a5637b16