Avoid masking exceptions while running hooks
Exceptions could be thrown by the 'Jack.check' or
'Jack.run' method, but since the 'runHooks' in the
finally block could throw an Exception as well, the first
exception could be masked.
To fix this, we run 'runHooks' first in the catch block
where the exception is rethrown. If there was no exception
we run 'runhooks', this time without catching exceptions.

Bug: 31239202
Change-Id: I002a1a77fc09fa2f774f7b53741694e795313621