Remove dependency on zip/unzip command line tools
This required to install those on Windows, on which it's
not present by default.
Jar has been used to zip files. It should be present because
Java is a requirement to run Jack tests on Windows.
Since jar can not unzip to a given directory, it's been implemented
by using the Java zip API.

Bug: 27642554

Change-Id: I31bcfdd513fd83571b58c80a99361fe415cf6c23