Add android.sdk property to test framework for DeviceRunner
Currently tests can be run on device/emulators provided that path to
SDK 'platform-tools' directory of Android SDK is in PATH env var.

Though this can usually be easy to set, this CL offers to
keep environment untouched by specifing the SDK location in the
test configuration file.

It also adds more explicit error messages in case adb could not
be found.

Change-Id: Ifb95a91a3863dbc1fa79c0a909e15045d06cff06