Protect code coverage plugin name
This CL ensures that the name of the plugin is not accidentally
changed after refactoring or repackaging. Such information is
now stored in a separate class.

Also renames the class CodeCoverage to CodeCoverageFeature since the
name of the plugin was its original name. Now the plugin name no
longer identifies any class.

Change-Id: I0c985c891a6cc492603bca9134ccc47f6ae3884f