Support code between def and use in IfWithConstantSimplifier
The IfWithConstantSimplifier only used to apply when there was no
code between the 'def' being optimized (the assign) and the 'use'
(the condition of the if statement). However, when code coverage is
enabled, it does insert extra instructions between def and use.
Therefore it prevents the optimization from being applied. In
presence of nested assignment in the 'if' statement, this will cause
ART verification failures at runtime.

This CL removes the constraint of "no code between def and use".
Because code coverage does not impact def/use chain here, it fixes
the ART verification failures when nested assignment are present.

Bug: 31759708
Test: 'ant test-jack' with api-v04
Change-Id: I322a91c55d9066d7520260d31a25d85e492db977