Use keys instead of Strings in HierarchyFinder
When searching for conflict between fields or methods in the hierarchy
(when renaming), we have to take the field type or method signature
into account.
To do this, we used to use strings "keys" and we were not doing any
check to enforce that these keys used the right format.
As a result, we used the FieldInHierarchyFinder with names instead of
keys which created a bug.
To fix it, I replaced the strings keys with a proper key object.

Bug: 32297886
Change-Id: If179bd940d92a20693a09d8b615bfc5b0d31c1ee