Adding 'type signature' to lambda group classes
With this CL we fix collisions in lamgda group class names in case
lambdas are coming from different compilations. This is relevant
in scenarios when both application and library were compiled
separately, as well as incremental compilation and some testing
scenarios.

Now lambda class name includes a signature of the type(s) defining
lambdas implemented in this group, thus binding the generated lambda
group class to the lambda creating classes.

Bug: 32366084
Bug: 32375858
Bug: 32268817
Change-Id: I6ab320d9dbe75f17600d0b81dbd36cc10254e8e1