Fix DictionaryNameProviderTest
The test was using the primitive type boolean which
is not possible in a unit test when the configuration
has not been initialized.
To fix this, I used a TypeKey instead of a FieldKey,
which only needs a name to be created. Note that the
DictionaryNameProvider uses the same code for FieldKeys
or TypeKeys.

Change-Id: I131ddbb463a938ac6675cde087ae3edcf7f5fb9c