First CL for Jack SSA IR
This is the first (of many) CL for Jack SSA IR. I feel this is a good time
to create a check point commit since it will not longer ask for big
features from the CFG-IR after this point.

Here is what is added.

1. jack.optimization.use-jack-ssa-ir property.
   This controls rather we transforms the CFG-IR right the SSA based one.

2. Forked SsaCodeItemBuilder that is exactly same as the original (for now).
   It is not completed yet so it will just assert when it finds a Phi Node.

3. Added JPhiBlockElement / JSsaVariableRef.
   That is something I mentioned in my previous presentation.

4. Turned most of the DominanceFrontier related things into a marker.

5. Use schedulables to compute all the D.F. related things.

6. SsaBasicBlockSplitter
   A schedulable to split basic blocks for Phi node insertion.

7. JPhiElementInsertion
   A schedulable to insert JPhiBlockElement

8. SsaRenamer
   A schedulable to rename all local variable references to JSsaVariableRef.

At the current state, the current CL has no behaviorial changes unless:
 -D jack.optimization.use-jack-ssa-ir=true

When it is on, it will insert Phi nodes, rename variables and crash
at CodeGen.

Change-Id: If32d58372b84874e3d2f210cb02e712f7c5d017b