Add call site section
- Header size section of the dex file varies according to dex version
- Update Jack merger to take into account call site, nevertheless the
index overflow of call site is not yet managed, it will be done in
a another CL.
- Add all items to represent a call site, CallSiteIdItem to represent
the content of the call site section. CstCallSiteRef to have a constant
representing a call-site and that will be used by invoke-custom instruction.
- Call sites index must not be internalized (uses appendIfAppropriate method)
even if the encoded array representing the data of a call site can be
internalized into the data section
- Extra constant arguments of linker method is not supported by the call site,
the support will be add later

Bug: 32361976

Change-Id: Iadd2e9d2ff02d4b10a0abe10de67528aa396b1c4