Fix IWCS when def and use are in the same block
The IfWithConstantSimplifier took the asumption that the definition
and its only use (the if statement) are in different blocks. Indeed
it is meant to optimize specific patterns built after transforming
'&&' and '||' expressions into conditional expressions.

This CL checks that the definition and the use are in different
blocks (actually JBlock). Otherwide we could generate incorrect code
by inserting a goto at a wrong position which would create an
infinite loop at runtime.

Bug: 33656257
Change-Id: I1b6eec9f8d2504083cc0bfd52a94e6db58ae9e77